Workaround linter issue
`PickleBuffer` isn't currently included in `SupportBytes`, which causes
issues with pyright when passing `PickleBuffer` instances to `bytes`.
Though it appears ok passing `PickleBuffer` instances to `memoryview`s.
So do that instead. This is functionaly very equivalent. There is a
slight performance cost to making a `memoryview`, but this is likely
negligible compared to copying to `bytes`.