Use `order="A"` in `memoryview.tobytes(...)`
In the function `flatten_buffer`, the order is already effectively
enforced when copying can be avoided by using `PickleBuffer.raw(...)`.
However some test comparisons failed (when they shouldn't have) as this
wasn't specified. So add the `order` in both the function and the test.
This should fix that test failure.