Add sheet to error message when reading Excel
When reading Excel files with `BaseExcelReader`, is possible that
`TextParser` raises `ValueError` on different situations. As the
reader works on a sheet basis, including the name of the sheet that
caused the exception can be useful for debugging.

The implemented solution captures any `ValueError` that gets raised
while parsing and reading a sheet from an Excel file, then adds the
name of the sheet to the exception message and "re-raises" the
exception using the `Exception.with_traceback` method.

Why use `with_traceback`? This allow-us to maintin the focus on the
actual line that raised the ValueError instead of directing the
attention to the introduced block of code.

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>