Test sheet name presence in `ValueError` message
The test performs a `read_excel` operation on a blank Excel file while
also passing `[1]` as header value. This operations results in a
`ValueError`.

Before the fix, the error message did not display any information
about the offending sheet, making it harder to debug Excel files with
several sheets. For this reason, the value of `sheet_name` was
explicitly set to its default value (`None`).

The test compares the `ValueError` message with the expected message
that contains the offending sheet name.

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>