PERF: Improve pickle support with BZ2 & LZMA (#49068)
* Add `BZ2File` wrapper for pickle protocol 5

* Add `LZMAFile` wrapper for pickle protocol 5

* Use BZ2 & LZMA wrappers for full pickle support

* Workaround linter issue

`PickleBuffer` isn't currently included in `SupportBytes`, which causes
issues with pyright when passing `PickleBuffer` instances to `bytes`.
Though it appears ok passing `PickleBuffer` instances to `memoryview`s.
So do that instead. This is functionaly very equivalent. There is a
slight performance cost to making a `memoryview`, but this is likely
negligible compared to copying to `bytes`.

* Refactor out `flatten_buffer`

* Refactor `B2File` into separate module

* Test `flatten_buffer`

This provides a reasonable proxy for testing patched `BZ2File` and
`LZMAFile` objects.

* Move `flatten_buffer` to `_utils`

This ran into cyclic import issues in `pickle_compat`. So move
`flatten_buffer` to its own module free of these issues.

* Import `annotations` to fix `|` usage

* Sort `import`s to fix lint

* Patch `BZ2File` & `LZMAFile` on Python pre-3.10

This should limit the effects of this patch. Also should make it easier
to remove this backport later once all supported Python versions have
the fix.

* Test C & F contiguous NumPy arrays

Also test another non-contiguous array.

* Test `memoryview` is 1-D `uint8` contiguous data

If a `memoryview` is returned, make sure it as close to `bytes` |
`bytearray` as possible. This ensures if other functions assume
something like `bytes` (for example assuming `len(b)` is the number of
bytes contained), things will continue to work even though this is a
`memoryview`.

* Run `black` on `bz2` and `lzma` compat files

* One more lint fix

* Drop unused `PickleBuffer` `import`s

* Simplify change to `panda.compat.__init__`

Now that the LZMA changes are in a separate file, cleanup the changes to
`pandas.compat.__init__`.

* Type `flatten_buffer` result

* Use `order="A"` in `memoryview.tobytes(...)`

In the function `flatten_buffer`, the order is already effectively
enforced when copying can be avoided by using `PickleBuffer.raw(...)`.
However some test comparisons failed (when they shouldn't have) as this
wasn't specified. So add the `order` in both the function and the test.
This should fix that test failure.

* Move all compat compressors into a single file

* Fix `BZ2File` `import`

* Refactor out common compat constants

* Fix `import` sorting

* Drop unused `import`

* Ignore `flake8` errors on wildcard `import`

* Revert "Ignore `flake8` errors on wildcard `import`"

This reverts commit f1f1a2e56861abbc0b2c04adbb98880794c47ca0.

* Explicitly `import` all constants

* Assign `IS64` first

* Try `noqa` on wildcard `import` again

* Declare `BZ2File` & `LZMAFile` once

Fixes a linter issue from pyright.

* `import PickleBuffer` for simplicity

* Add `bytearray` to return type

* Test `bytes` & `bytearray` are returned unaltered

* Explicit list all constants

* Trick linter into thinking constants are used ;)

* Add new entry to 2.0.0

* Assign constants to themselves

Should work around linter issues.

* Update changelog entry    [skip ci]

* Add constants to `__all__`

* Update changelog entry    [ci skip]

* Use Sphinx method annotation