Add sheet name to any Exception traceback
Instead of including the sheet name on the traceback when a ValueError
is raised, we'll do it for any kind of Exception. In particular, this
is because some of the arguments for `read_excel` are callables, and
they can raise any Exception. Including the sheet name might be
helpful in some of those cases, so we'll capture them too

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>