parent 0168e27843efb96b56202f22fbe5dd720c346368 author Antonio Ossa Guerra <aaossa@uc.cl> 1666800993 -0300 committer MarcoGorelli <> 1669293453 +0000
Parse `datetime` properly in `pd.to_datetime`

When applying `pd.to_datetime` on array-like structure that contain a
`datetime.datetime` object, while using the `format` argument, a
`ValueError` is raised because the `datetime.datetime` object does not
match the expected format.

The implemented solution looks for `datetime.datetime` instances in the
`array_strptime` method. If an instance of this type is found, it's
properly handled by the new `_parse_python_datetime_object`, which
returns the expected Numpy datetime object.

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>