Test `pd_to_datetime` in presence of `datetime`
Two tests are introduced to check that `datetime.datetime` objects are
skipped while parsing array-like structures to `pd.to_datetime` and
the `format` argument is being used.

Before the fix, all of these tests and their test cases failed, except
`test_to_datetime_arraylike_contains_pydatetime_utc` when cache was
`True` and the `init_constructor` was `Index`.

The test creates an input of data that must be formated and a
`datetime.datetime` object. Then, the values are passed to the
array-like `init_constructor` to be converted to datetime using the
`pd.to_datetime` method. The test must not fail, and the result must
match the expected data.

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>