Return localized object if a timezone is present
When using the `format` argument on `pd.to_timestamp`, and the input
contains localized `pd.Timestamp` or `datetime.datetime` objects, the
result should be localized too.

A bug was triggered when `format` did not contain `%z` or `%Z`,
because the `_array_strptime_with_fallback` function did not expect
any type different than string. Now this assumption is no longer
valid, since we now let `pd.Timestamp` and `datetime.datetime` objects
pass unmodified.

The fix is to change the condition that checks if returning a
localized output is necessary: instead of looking at the passed
`format`, return a localized result when any input object contains a
timezone.

Also, a new test is added to check this case. A case is included for a
`pd.Timestamp` and another for a `datetime.datetime` object, and each
compares the actual result with the expected result when the input is
localized and when is not.

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>