Raise `ValueError` when tz-aware differs in input
Just like `to_datetime` when `format` is not passed, we should raise a
`ValueError` if the input contains some tz-aware objects and some
naive objects. The change is to identify if the `format` expects
tz-aware objects (check for `"%z"` and `"%Z"`) and then compare this
expectation with the `tzinfo` of each object. This modification
includes relevant tests.

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>