Update holiday.py
Added a check in the dates method to address the bug mentioned in issue #49925. If the dates method is ran for a date range outside of the holiday's start_date/end_date the holiday_dates object is converted from Index to DateTimeIndex and daysofweek is no longer invoked on an Index object.