Update generic.py
This is a prototype for how `groupby` could create extension arrays from homogeneous elements that can be made into EAs.  It is an attempt to address https://github.com/hgrecco/pint-pandas/issues/142

The prototype exposes the fact that we don't have an easy way of connecting the dots from q Quantity (defined in Pint) and a PintArray (defined in Pint-Pandas).  But if we did have such a thing, then dataframes of heterogeneous units (such as quantities of production -- tons of Steel, tons of Aluminum, TWh, etc) could be grouped into homogeneous units presented in their natural EA types.

Signed-off-by: Michael Tiemann <72577720+MichaelTiemannOSC@users.noreply.github.com>