ENH: use shutil.which() instead of external which(1) (#54937)
* ENH: update bundled pyperclip with changes from 1.8.2 release

Copy the changes from upstream 1.8.2 to the bundled copy of pyperclip.
The code was reformatted using black and verified using ruff.
The existing modifications from pandas were preserved.

* ENH: Remove Python 2 compatibility from imported pyperclip code

Remove the fallback to which/where that is only necessary for Python 2
that does not feature shutil.which().  Also collapse the imports
to avoid importing shutil.which() twice.  It is now only imported
as `_executable_exists()` to minimize the changes to the original code.

* BUG: Fix pylint failure (redundant `pass`) in clipboard