BLD: Allow building with NumPy nightlies and use it for nightlies
I am not sure that this is uncontested that this is a good version
pin change.  I also took the liberty of removing the current use of
oldest-support-numpy, since Python 3.9 means that you should normally
always get 1.25 anyway and that should work for any NumPy version.

The `<=2.0.0.dev0` is a way to explicitly allow NumPy nightly wheels,
although of course you need the additional URL to use it.
In a release version, it might be nicer to not have it, but it also
should not really hurt since the you are unlikely to have the NumPy dev
wheels available unless you really want them.

(Not sure how to best test that the changes work as is.)