immitating interactive mode
For each S3 bucket code block, ideally we show what the output would be, but without making an actual call. Unfortunately, for several of the S3 buckets, there are issues with the code, which we must fix in another commit or PR.

For now, the two S3 examples that do work, we edit to make the code block show what the output would have been if it had run successfully.

Find details on issues in conversation on PR #56592