DOC: Replace pandas logo with a conditional image for light and dark … (#57223)
DOC: Replace pandas logo with a conditional image for light and dark themes

Replaced the static pandas logo with a `picture` tag that displays a different logo based on the color scheme. In light mode, the original logo is shown, while in dark mode, a white version of the logo is displayed for better visibility.