Fix evaluations on Python 3.12
List comprehensions no longer get their own scope [1], so adding a level
to the (eventual) call to `sys._getframe` goes outside the actual
caller. If running in `pytest` (so that there is a scope outside the
caller), you end up looking in some unrelated scope. If you are running
a script, then `sys._getframe` raises an error that the level is out of
bounds.

The `Bitwise operations` warning in `test_scalar_unary` appears to
always be raised, so remove the condition.

[1] https://docs.python.org/3.12/whatsnew/3.12.html#whatsnew312-pep709