Allow @np to be used within df.eval and df.query
Add additional check to make sure that ndim of the provided
variable is an int. This makes sure that the ndim guard doesn't
trigger if it is something else (e.g., a function in case of a
numpy). This allow for using @np in df.eval and df.query methods.