BUG: Allow using numpy in `DataFrame.eval` and `DataFrame.query` via @-notation (#58057)
* Allow @np to be used within df.eval and df.query

Add additional check to make sure that ndim of the provided
variable is an int. This makes sure that the ndim guard doesn't
trigger if it is something else (e.g., a function in case of a
numpy). This allow for using @np in df.eval and df.query methods.

* Add whatsnew

* Fix typo

Co-authored-by: Abdulaziz Aloqeely <52792999+Aloqeely@users.noreply.github.com>

* Test: skip checking names due to inconsistencies between OSes

* Elaborate futher on whatsnew message

* Fix the test by explicitly specifing engine.

Also move the test to more appropriate file.

---------

Co-authored-by: Abdulaziz Aloqeely <52792999+Aloqeely@users.noreply.github.com>