ENH: Support Greek small letter mu
Characters which look the same but have different underlying codes (homographs) are problematic. The abbreviation for microseconds is best represented as μs, using the 0x3bc Greek small letter mu, instead of the 0xb5 Latin Extended micro sign. This aligns with:

- [The International System of Units (SI) brochure](https://www.bipm.org/documents/20126/41483022/SI-Brochure-9-EN.pdf)
- NFKC normalized [Python code](https://peps.python.org/pep-3131/) and [domain names](https://unicode.org/reports/tr36/)
- Section 2.5 Duplicated Characters of [Unicode Technical Report 25](https://www.unicode.org/reports/tr25/)
- The microfarads abbreviation in [Pandas tests](/pandas-dev/pandas/tree/2.2.x/pandas/tests/computation/test_eval.py#L1914)
- Ruff confusable mapping [updates](https://github.com/astral-sh/ruff/pull/4430/files) (currently in the "preview" stage)

Ensure that both Latin Extended and Greek characters are accepted as aliases for microseconds. Prefer the Greek character in a comment and UTF8 test data.