Add choice of index source internally
This commit introduces an index argument that tells merging how
to choose the output index. I moved the logic determining which
index to use into _MergeOperation.__init__ while moving the logic
for calculating the index to the end of the _get_join_info method.
Note: I did not adjust the logic in the case when merging on both
indexes (or an index and a multiindex).

This commit also introduces a function merge_pick_index so that
internal uses of merge can pick where they get the index from.
Specifically, join uses merge under the hood, but wants the old
merge index sources. Now, join specifies the index source so it
wont be affected by changing merge's index behavior.

This commit should not affect behavior.