Update .gitpod.yml
Updated the tasks when setting up the Gitpod environment. All command line output during setup should now be logged into a file (setup.log), and echos are done before each segment of setup. This way, we should be able to narrow down the exact command that causes the weird pandas version to be downloaded.