BUG: Fix DataFrame constructor misclassification of array-like with 'name' attribute (#61443) (#61451)
* BUG: Fix DataFrame constructor misclassification of array-like with 'name' attribute

Previously, any object with a .name attribute (like some vtkArray-like objects) was assumed to be a Series or Index, causing the constructor to misinterpret the input.

This fix ensures we only apply the named-Index/Series logic when the input is actually an instance of ABCSeries or ABCIndex *and* has a non-None name.

Closes #61443.

* Apply pre-commit fixes: isort and remove unused type ignore

* TST/CLN: Add issue reference in test and update whatsnew for GH#61443