BUG: Copy attrs on pd.merge() (#60357)
* BUG: Copy attrs on pd.merge()

This uses the same logic as `pd.concat()`: Copy `attrs` only if all
input `attrs` are identical.

I've refactored the handling in __finalize__ from special-casing based on th the method name (previously only "concat") to handling "other" parameters
that have an `input_objs` attribute. This is a more scalable architecture compared to hard-coding method names in __finalize__.

Tests added for `concat()` and `merge()`.

Closes #60351.

* Update docs

* Add release note

* Fix wrong link

Co-authored-by: Matthew Roeschke <10647082+mroeschke@users.noreply.github.com>

* Move release note to v3.0.0

---------

Co-authored-by: Matthew Roeschke <10647082+mroeschke@users.noreply.github.com>
Co-authored-by: Marc Garcia <garcia.marc@gmail.com>